+++ models/sdatabaserecord.cpp
@@ -144,8 +144,8 @@ QString SDatabaseRecord::fieldValueHandler(const QVariant &value)
         str_value =  value.toDate().toString("yyyy-MM-dd");
         str_value =  "'" + str_value + "'";
     }
-    else if(value.typeName() == QString("float"))
-        str_value =  QString::number(value.toFloat(), 'f', 4);
+    else if(value.typeName() == QString("double"))
+        str_value =  QString::number(value.toDouble(), 'f', 4);
     else if(value.typeName() == QString("double"))
         str_value =  QString::number(value.toDouble(), 'f', 4);
     else
